+++ src/Executors/DatabaseRawQueryExecutor.php
@@ -14,9 +14,9 @@ class DatabaseRawQueryExecutor extends AbstractExecutor
     /**
      * {@inheritdoc}
      */
-    public function execute($data, string $connection_name = null): bool
+    public function execute($data, ?string $connection_name = null): bool
     {
-        if (\is_string($data) && ! empty($data)) {
+        if (\is_string($data) && $data !== '') {
             /** @var ConnectionInterface $connection */
             $connection = $this->app->make('db')->connection($connection_name);