+++ tests/Commands/AbstractCommandTestCase.php
@@ -18,7 +18,7 @@ abstract class AbstractCommandTestCase extends AbstractTestCase
      *
      * @return void
      */
-    public function testHelpCommand()
+    public function testHelpCommand(): void
     {
         $this->assertNotFalse(
             $this->artisan($signature = $this->getCommandSignature(), ['--help']),
@@ -31,5 +31,5 @@ abstract class AbstractCommandTestCase extends AbstractTestCase
      *
      * @return string
      */
-    abstract protected function getCommandSignature();
+    abstract protected function getCommandSignature(): string;
 }