+++ tests/Traits/ApplicationHelpersTrait.php
@@ -7,10 +7,6 @@ use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Illuminate\Contracts\Foundation\Application as ApplicationContract;
 
 /**
- * Class ApplicationHelpersTrait.
- *
- * Трейт вспомогательных методов по работе с инстансом приложения и подобные.
- *
  * @mixin \Illuminate\Foundation\Testing\TestCase
  */
 trait ApplicationHelpersTrait
@@ -38,7 +34,7 @@ trait ApplicationHelpersTrait
      *
      * @return Kernel|\App\Console\Kernel
      */
-    public function console(ApplicationContract $app = null)
+    public function console(?ApplicationContract $app = null)
     {
         $app = $this->resolveApplication($app);
 
@@ -52,7 +48,7 @@ trait ApplicationHelpersTrait
      *
      * @return ConfigRepository
      */
-    public function config(ApplicationContract $app = null)
+    public function config(?ApplicationContract $app = null)
     {
         $app = $this->resolveApplication($app);
 
@@ -82,7 +78,7 @@ trait ApplicationHelpersTrait
      *
      * @return void
      */
-    public function setAppEnvironment($environment = 'testing', ApplicationContract $app = null)
+    public function setAppEnvironment($environment = 'testing', ?ApplicationContract $app = null)
     {
         $app = $this->resolveApplication($app);
 
@@ -97,7 +93,7 @@ trait ApplicationHelpersTrait
      * @param bool                     $recreate_db
      * @param ApplicationContract|null $app
      */
-    public function prepareDatabase($recreate_db = true, ApplicationContract $app = null)
+    public function prepareDatabase($recreate_db = true, ?ApplicationContract $app = null)
     {
         $app = $this->resolveApplication($app);
 
@@ -135,7 +131,7 @@ trait ApplicationHelpersTrait
      *
      * @return void
      */
-    public function clearCache(ApplicationContract $app = null)
+    public function clearCache(?ApplicationContract $app = null)
     {
         $this->console($app)->call('cache:clear');
     }
@@ -147,7 +143,7 @@ trait ApplicationHelpersTrait
      *
      * @return ApplicationContract|\Illuminate\Foundation\Application
      */
-    protected function resolveApplication(ApplicationContract $app = null)
+    protected function resolveApplication(?ApplicationContract $app = null)
     {
         if ($app instanceof ApplicationContract) {
             return $app;